settings and rename
