+++ examples/example-typescript/.codegraphy/settings.json
@@ -1,6 +1,7 @@
 {
   "version": 1,
   "maxFiles": 1000,
+  "verboseDiagnostics": false,
   "include": [
     "**/*"
   ],
@@ -46,15 +47,15 @@
     "plugin:codegraphy.gdscript:symbol:godot-class-name": false
   },
   "edgeVisibility": {
-    "nests": false,
     "import": true,
-    "type-import": false,
-    "reexport": false,
-    "call": false,
-    "inherit": false,
     "reference": false,
+    "call": false,
     "test": false,
+    "reexport": false,
+    "type-import": false,
+    "inherit": false,
     "load": false,
+    "nests": false,
     "contains": false,
     "overrides": false
   },
