+++ examples/example-vue/src/components/StatusBadge.vue
