+++ examples/example-vue/src/components/UserCard.vue
