+++ examples/example-vue/src/types.ts
