+++ examples/example-vue/src/vue.d.ts
