+++ examples/example-vue/vite.config.ts
