+++ packages/extension/tests/acceptance/specs/go-example.md
@@ -9,7 +9,7 @@ Then I see graph nodes
 And I see edges
 And the graph nodes match the expected files in the examples/example-go workspace
 
-Then I can see there are 5 nodes and 2 connection
+Then I can see there are 5 nodes and 1 connection
 And main.go points to internal/service/service.go
 
 And README.md is an orphan node
