+++ packages/extension/tests/acceptance/specs/java-example.md
@@ -9,7 +9,7 @@ Then I see graph nodes
 And I see edges
 And the graph nodes match the expected files in the examples/example-java workspace
 
-Then I can see there are 5 nodes and 2 connections
+Then I can see there are 5 nodes and 1 connections
 And src/com/example/app/App.java points to src/com/example/app/Helper.java
 
 And README.md is an orphan node
