+++ packages/extension/tests/acceptance/specs/rust-example.md
@@ -9,7 +9,7 @@ Then I see graph nodes
 And I see edges
 And the graph nodes match the expected files in the examples/example-rust workspace
 
-Then I can see there are 6 nodes and 3 connections
+Then I can see there are 6 nodes and 2 connections
 And src/main.rs points to src/util.rs
 And src/main.rs points to src/inner.rs
