+++ packages/extension/tests/acceptance/specs/typescript-example.md
@@ -9,7 +9,7 @@ Then I see graph nodes
 And I see edges
 And the graph nodes match the expected files in the examples/example-typescript workspace
 
-Then I can see there are 12 nodes and 5 connections
+Then I can see there are 11 nodes and 5 connections
 And src/index.ts points to src/types.ts
 And src/index.ts points to src/utils.ts
 And src/utils.ts points to src/depth.ts
@@ -21,11 +21,10 @@ And package.json is an orphan node
 And .gitignore is an orphan node
 And tsconfig.json is an orphan node
 And src/alias/greeting.ts is an orphan node
-And .vscode/settings.json is an orphan node
 
 When I click the plugins button
 Then I see a list of plugins with toggles
 And I toggle the TypeScript/JavaScript plugin on
-Then I can see there are 12 nodes and 6 connections
+Then I can see there are 11 nodes and 6 connections
 And I see src/index.ts points to src/alias/greeting.ts
 And src/alias/greeting.ts is no longer an orphan node
